backend/requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.7.1
diff-match-patch==20200713
django-filter==2.4.0
django-import-export==2.5.0
et-xmlfile==1.0.1
idna==2.9
importlib-metadata==1.6.0
MarkupPy==1.14
more-itertools==8.2.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
PyYAML==5.4.1
requests==2.23.0
six==1.14.0
tablib==3.0.0
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.1.9
xlrd==2.0.1
xlwt==1.3.0
zipp==3.1.0
django-filter==2.4.0
flake8
pylint
timestamp~=0.0.1
django-import-export
djoser
djangorestframework_simplejwt


asgiref==3.2.10
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1 
PyPI
psycopg2==2.8.6
psycopg2-binary==2.8.5